Move source files and add Doxygen labels
