+++ TESTS/mbedmicro-rtos-mbed/systimer/main.cpp
@@ -23,7 +23,7 @@
 extern "C" {
 #include "rtx_lib.h"
 }
-#include "platform/SysTimer.h"
+#include "platform/source/SysTimer.h"
 
 #define TEST_TICKS 42
 #define TEST_TICK_US (TEST_TICKS * 1000)
