+++ rtos/source/Kernel.cpp
@@ -24,7 +24,7 @@
 #include "rtos_idle.h"
 #include "rtos_handlers.h"
 #include "platform/mbed_critical.h"
-#include "platform/mbed_os_timer.h"
+#include "platform/source/mbed_os_timer.h"
 
 #if !MBED_CONF_RTOS_PRESENT
 /* If the RTOS is not present, we call mbed_thread.cpp to do the work */
